+++ plugins/inputs/system/memory.go
@@ -30,6 +30,8 @@ func (s *MemStats) Gather(acc telegraf.Accumulator) error {
 		"free":              vm.Free,
 		"cached":            vm.Cached,
 		"buffered":          vm.Buffers,
+		"active":            vm.Active,
+		"inactive":          vm.Inactive,
 		"used_percent":      100 * float64(vm.Used) / float64(vm.Total),
 		"available_percent": 100 * float64(vm.Available) / float64(vm.Total),
 	}